1192: missed renaming bits
