+++ web-client/src/applicationContext.js
@@ -84,7 +84,7 @@ import { getProcedureTypes } from '../../shared/src/business/useCases/getProcedu
 import { getScannerInterface } from '../../shared/src/business/useCases/getScannerInterfaceInteractor';
 import { getSentWorkItemsForSection } from '../../shared/src/proxies/workitems/getSentWorkItemsForSectionProxy';
 import { getSentWorkItemsForUser } from '../../shared/src/proxies/workitems/getSentWorkItemsForUserProxy';
-import { getTrialSessionDetails } from '../../shared/src/business/useCases/trialSessions/getSessionDetailsInteractor';
+import { getTrialSessionDetails } from '../../shared/src/business/useCases/trialSessions/getTrialSessionDetailsInteractor';
 import { getTrialSessions } from '../../shared/src/proxies/trialSessions/getTrialSessionsProxy';
 import { getUser } from '../../shared/src/business/useCases/getUserInteractor';
 import { getUsersInSection } from '../../shared/src/proxies/users/getUsersInSectionProxy';
